\name{kf}
\alias{kf}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Kernel function
}
\description{
Kernel function
}
\usage{
kf(x, h)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
x
}
  \item{h}{
bandwidth
}
}
\value{Returns value for the kernel at the given x}
\keyword{internal}
