\name{mu0wy}
\alias{mu0wy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimates mu0(w)=E(Y0|W0=w)
}
\description{
Estimates mu0(w)=E(Y0|W0=w)
}
\usage{
mu0wy(w, hyw0,y0,w0)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{w}{
w
}
  \item{hyw0}{
bandwidth
}
	 \item{y0}{
y0
}
	 \item{w0}{
w0
}
}
\value{Returns mu0(w) estimate for given w}
\keyword{internal}
