\name{mu10}
\alias{mu10}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimates mu10(w) 
}
\description{
Estimates mu10(w) which is the integral of mu1(s, w) with respect to dF0(s|w)
}
\usage{
mu10(s1,w1,y1,w0,s0,w, hy2w1, hy2s1, hsw1)
}
\arguments{
  \item{s1}{
s1
}
  \item{w1}{
w1
}
  \item{y1}{
y1
}
  \item{w0}{
w0
}
  \item{s0}{
s0
}
  \item{w}{
w
}
  \item{hy2w1}{
bandwidth
}
  \item{hy2s1}{
bandwidth
}
  \item{hsw1}{
bandwidth
}
}
\value{Returns mu10(w) estimate for given w}
 \keyword{internal}

