\name{mu1swy}
\alias{mu1swy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimates mu1(s, w)=E(Y1|S1=s, W1=w)
}
\description{
Estimates mu1(s, w)=E(Y1|S1=s, W1=w)
}
\usage{
mu1swy(s1,w1,y1,s, w, hy2w1, hy2s1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{s1}{
s1
}
  \item{w1}{
w1
}
  \item{y1}{
y1
}
  \item{s}{
s
}
  \item{w}{
w
}
  \item{hy2w1}{
bandwidth
}
  \item{hy2s1}{
bandwidth
}
}
\value{Returns mu1(s,w) estimate for given s and w}
\keyword{internal}