\name{mu1wy}
\alias{mu1wy}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Estimates mu1(w)=E(Y1|W1=w)
}
\description{
Estimates mu1(w)=E(Y1|W1=w)
}
\usage{
mu1wy(w, hyw1,y1,w1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{w}{
w
}
  \item{hyw1}{
bandwidth
}
  \item{y1}{
y1
}
  \item{w1}{
w1
}
}
\value{Returns mu1(w) estimate for given w}
 \keyword{internal}
